Ajout modification formuleraire USER/ADMIN

+++ src/Controller/Admin/UserController.php
@@ -4,100 +4,92 @@ namespace App\Controller\Admin;
 
 use App\Entity\User;
 use App\Form\UserType;
+use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
 #[Route('/admin/user')]
 class UserController extends AbstractController
 {
-    #[Route('/', name: 'admin_user_index', methods: ['GET'])]
-    public function index(EntityManagerInterface $em): Response
+    #[Route('/', name: 'app_admin_user_index', methods: ['GET'])]
+    public function index(UserRepository $userRepository): Response
     {
-        dump($this->getUser()->getRoles());
-
-        die();
-
-        $users = $em->getRepository(User::class)->findAll();
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
         return $this->render('admin/user/index.html.twig', [
-            'users' => $users,
+            'users' => $userRepository->findAll(),
         ]);
     }
 
-    #[Route('/new', name: 'admin_user_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $em, UserPasswordHasherInterface $passwordHasher): Response
+    #[Route('/new', name: 'app_admin_user_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $em, UserPasswordHasherInterface $hasher): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $user = new User();
         $form = $this->createForm(UserType::class, $user, ['is_admin' => true]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $plainPassword = $form->get('password')->getData();
+            $hashedPassword = $hasher->hashPassword($user, $user->getPassword());
+            $user->setPassword($hashedPassword);
 
-            dump($form->get('password')->getData());
+            $em->persist($user);
+            $em->flush();
 
-            if (!$plainPassword) {
-                $form->get('password')->addError(new FormError('Le mot de passe ne peut pas être vide.'));
-            } else {
-                $hashedPassword = $passwordHasher->hashPassword($user, $plainPassword);
-                $user->setPassword($hashedPassword);
+            $this->addFlash('success', 'Utilisateur créé avec succès.');
 
-                $em->persist($user);
-                $em->flush();
-
-                return $this->redirectToRoute('admin_user_index');
-            }
+            return $this->redirectToRoute('app_admin_user_index');
         }
 
         return $this->render('admin/user/new.html.twig', [
-            'userForm' => $form->createView(),
-        ]);
-    }
-
-    #[Route('/{id}', name: 'admin_user_show', methods: ['GET'])]
-    public function show(User $user): Response
-    {
-        return $this->render('admin/user/show.html.twig', [
             'user' => $user,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'admin_user_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, User $user, EntityManagerInterface $em, UserPasswordHasherInterface $passwordHasher): Response
+    #[Route('/{id}/edit', name: 'app_admin_user_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, User $user, EntityManagerInterface $em, UserPasswordHasherInterface $hasher): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(UserType::class, $user, ['is_edit' => true, 'is_admin' => true]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $plainPassword = $form->get('password')->getData();
             if ($plainPassword) {
-                $hashedPassword = $passwordHasher->hashPassword($user, $plainPassword);
+                $hashedPassword = $hasher->hashPassword($user, $plainPassword);
                 $user->setPassword($hashedPassword);
             }
-            // Sinon on garde le mot de passe actuel
 
             $em->flush();
 
-            return $this->redirectToRoute('admin_user_index');
+            $this->addFlash('success', 'Utilisateur mis à jour.');
+            return $this->redirectToRoute('app_admin_user_index');
         }
 
         return $this->render('admin/user/edit.html.twig', [
-            'userForm' => $form->createView(),
+            'user' => $user,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/{id}', name: 'admin_user_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_admin_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $em): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $em->remove($user);
             $em->flush();
+            $this->addFlash('success', 'Utilisateur supprimé.');
         }
 
-        return $this->redirectToRoute('admin_user_index');
+        return $this->redirectToRoute('app_admin_user_index');
     }
 }
